Add a script to build and compile heron in a container
This adds two separate environment for which heron can be compiled in,
ubuntu 14.04 and centos 7.

An example usage of the script would look like:
`docker/build-artifacts.sh ubuntu 0.1.0-SNAPSHOT ./docker`

This will compile heron using the ubuntu docker container, with the
version identifier 0.1.0-SNAPSHOT and put the resulting artifacts into
the ./docker folder. The output folder can be anywhere but it is worth
noting that on OS X the folder must be somewhere under the /Users
directory in order for virtual box to properly mount the founder and
transfer the artifacts.