Export heron-spi jar (#937)
* Export heron-spi jar

The heron-spi jar is needed for developing custom schedulers, packing
algorithms, etc. We will want to upload this jar to maven central along
with the api and storm compatibility layer.

* Add generate spi jar to api install script

We want the spi jar to be installed along with the api and storm
compatibility jars so that it is available for users to write code
against the interfaces.